pre testing
