+++ server/api/crud/model_monitoring/deployment.py
@@ -29,6 +29,7 @@ import mlrun.model_monitoring.stream_processing
 import mlrun.model_monitoring.writer
 import mlrun.serving.states
 import server.api.api.endpoints.functions
+import server.api.api.endpoints.nuclio
 import server.api.api.utils
 import server.api.crud.model_monitoring.helpers
 import server.api.utils.scheduler
@@ -230,6 +231,67 @@ class MonitoringDeployment:
             "writer_ready": ready,
         }
 
+    def apply_and_create_stream_trigger(
+        self, function: mlrun.runtimes.ServingRuntime, function_name: str = None
+    ) -> mlrun.runtimes.ServingRuntime:
+        """Adding stream source for the nuclio serving function. By default, the function has HTTP stream trigger along
+        with another supported stream source that can be either Kafka or V3IO, depends on the stream path schema that is
+        defined under mlrun.mlconf.model_endpoint_monitoring.store_prefixes. Note that if no valid stream path has been
+        provided then the function will have a single HTTP stream source.
+
+        :param function:                    The serving function object that will be applied with the stream trigger.
+        :param function_name:               The name of the function that be applied with the stream trigger,
+                                            None for model_monitoring_stream
+
+        :return: ServingRuntime object with stream trigger.
+        """
+
+        # Get the stream path from the configuration
+        # stream_path = mlrun.mlconf.get_file_target_path(project=project, kind="stream", target="stream")
+        stream_path = server.api.crud.model_monitoring.get_stream_path(
+            project=self.project, function_name=function_name
+        )
+
+        if stream_path.startswith("kafka://"):
+            topic, brokers = mlrun.datastore.utils.parse_kafka_url(url=stream_path)
+            # Generate Kafka stream source
+            stream_source = mlrun.datastore.sources.KafkaSource(
+                brokers=brokers,
+                topics=[topic],
+            )
+            function = stream_source.add_nuclio_trigger(function)
+
+        if not mlrun.mlconf.is_ce_mode():
+            function = self._apply_access_key_and_mount_function(
+                function=function, function_name=function_name
+            )
+            if stream_path.startswith("v3io://"):
+                server.api.api.endpoints.nuclio.create_model_monitoring_stream(
+                    project=self.project,
+                    monitoring_application=function_name
+                    != mm_constants.MonitoringFunctionNames.STREAM,
+                    stream_path=stream_path,
+                    access_key=self.model_monitoring_access_key,
+                )
+                kwargs = {}
+                if function_name != mm_constants.MonitoringFunctionNames.STREAM:
+                    kwargs["access_key"] = self.model_monitoring_access_key
+                if mlrun.mlconf.is_explicit_ack(version=resolve_nuclio_version()):
+                    kwargs["explicit_ack_mode"] = "explicitOnly"
+                    kwargs["worker_allocation_mode"] = "static"
+
+                # Generate V3IO stream trigger
+                function.add_v3io_stream_trigger(
+                    stream_path=stream_path,
+                    name=f"monitoring_{function_name or 'stream'}_trigger",
+                    **kwargs,
+                )
+        # Add the default HTTP source
+        http_source = mlrun.datastore.sources.HttpSource()
+        function = http_source.add_nuclio_trigger(function)
+
+        return function
+
     def _initial_model_monitoring_stream_processing_function(
         self,
         stream_image: str,
@@ -276,7 +338,7 @@ class MonitoringDeployment:
         function.metadata.project = self.project
 
         # Add stream triggers
-        function = self._apply_and_create_stream_trigger(
+        function = self.apply_and_create_stream_trigger(
             function=function, function_name=mm_constants.MonitoringFunctionNames.STREAM
         )
 
@@ -319,67 +381,6 @@ class MonitoringDeployment:
         server.api.api.utils.apply_enrichment_and_validation_on_function(
             function, self.auth_info
         )
-
-        return function
-
-    def _apply_and_create_stream_trigger(
-        self, function: mlrun.runtimes.ServingRuntime, function_name: str = None
-    ) -> mlrun.runtimes.ServingRuntime:
-        """Adding stream source for the nuclio serving function. By default, the function has HTTP stream trigger along
-        with another supported stream source that can be either Kafka or V3IO, depends on the stream path schema that is
-        defined under mlrun.mlconf.model_endpoint_monitoring.store_prefixes. Note that if no valid stream path has been
-        provided then the function will have a single HTTP stream source.
-
-        :param function:                    The serving function object that will be applied with the stream trigger.
-        :param function_name:               The name of the function that be applied with the stream trigger,
-                                            None for model_monitoring_stream
-
-        :return: ServingRuntime object with stream trigger.
-        """
-
-        # Get the stream path from the configuration
-        # stream_path = mlrun.mlconf.get_file_target_path(project=project, kind="stream", target="stream")
-        stream_path = server.api.crud.model_monitoring.get_stream_path(
-            project=self.project, function_name=function_name
-        )
-
-        if stream_path.startswith("kafka://"):
-            topic, brokers = mlrun.datastore.utils.parse_kafka_url(url=stream_path)
-            # Generate Kafka stream source
-            stream_source = mlrun.datastore.sources.KafkaSource(
-                brokers=brokers,
-                topics=[topic],
-            )
-            function = stream_source.add_nuclio_trigger(function)
-
-        if not mlrun.mlconf.is_ce_mode():
-            function = self._apply_access_key_and_mount_function(
-                function=function, function_name=function_name
-            )
-            if stream_path.startswith("v3io://"):
-                server.api.api.endpoints.functions.create_model_monitoring_stream(
-                    project=self.project,
-                    monitoring_application=function_name
-                    != mm_constants.MonitoringFunctionNames.STREAM,
-                    stream_path=stream_path,
-                    access_key=self.model_monitoring_access_key,
-                )
-                kwargs = {}
-                if function_name != mm_constants.MonitoringFunctionNames.STREAM:
-                    kwargs["access_key"] = self.model_monitoring_access_key
-                if mlrun.mlconf.is_explicit_ack(version=resolve_nuclio_version()):
-                    kwargs["explicit_ack_mode"] = "explicitOnly"
-                    kwargs["worker_allocation_mode"] = "static"
-
-                # Generate V3IO stream trigger
-                function.add_v3io_stream_trigger(
-                    stream_path=stream_path,
-                    name=f"monitoring_{function_name or 'stream'}_trigger",
-                    **kwargs,
-                )
-        # Add the default HTTP source
-        http_source = mlrun.datastore.sources.HttpSource()
-        function = http_source.add_nuclio_trigger(function)
 
         return function
 
@@ -445,7 +446,7 @@ class MonitoringDeployment:
         function.metadata.project = self.project
 
         # Add stream triggers
-        function = self._apply_and_create_stream_trigger(
+        function = self.apply_and_create_stream_trigger(
             function=function,
             function_name=mm_constants.MonitoringFunctionNames.WRITER,
         )
